starting fresh

+++ fresh.d.post/tmuxp.sh
@@ -1,2 +1,4 @@
-rm -f ~/.tmuxp
-ln -s ~/.dotfiles/tmuxp ~/.tmuxp
+if command -v tmuxp >/dev/null 2>&1; then
+  rm -f ~/.tmuxp
+  ln -s ~/.dotfiles/tmuxp ~/.tmuxp
+fi
